Fixed recording of query.
When running with concurrent streams, not all queries were recorded
successfully to the database like MySQL due to the timeToExit signal
channel being received first which would prematurely end the loop
that records the query to the database even though that resultChan
is not done receiving all the queries report yet.

Fix was to close the resultChan and modify the loop to allow for full
processing of the resultChan even if the timeToExit signal channel was
received.